Project modifications - fixed size_t to int cast warnings - fixed include paths for libusb liblo - removed some lua binding methods that cause build errors (need to check them again to see if they should be added back) - removed LibraryEditor (agaiN!)

+++ Source/Core/CtrlrSysexProcessor.cpp
@@ -147,7 +147,7 @@ CtrlrMidiMessageEx CtrlrSysexProcessor::sysexMessageFromString (const String &fo
 	Array <CtrlrSysexToken> tokenArray = sysExToTokenArray(formula);
 
 	CtrlrMidiMessageEx mex;
-	mex.m  = MidiMessage (bl.getData(), bl.getSize());
+	mex.m  = MidiMessage (bl.getData(), (int)bl.getSize());
 	mex.setTokenArray(tokenArray);
 	return (mex);
 }
